Archive/requirements.txt
# YouTube Analytics Pipeline with Transcription & Analysis
# Complete end-to-end pipeline for video data collection, transcription, and content analysis

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Video/Audio processing
yt-dlp>=2023.12.0

# Transcription (Whisper)
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Content Analysis (OpenAI)
openai>=1.0.0

# Optional: GPU support (uncomment if using NVIDIA GPU)
# torch==2.0.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

# Note: Requires ffmpeg to be installed separately
# macOS: brew install ffmpeg
# Ubuntu: sudo apt install ffmpeg
# Windows: Download from https://ffmpeg.org/
